capstone-project-2025-t1-25t1-9900-f14a-Avocado-main/ai_agent/requirements.txt
Flask>=3.1.0
flask-cors>=5.0.1
python-dotenv>=1.0.1
supabase>=2.15.0
Flask-SocketIO>=5.5.1
openai>=1.76.0